refactor: move component logic to custom hook
